HACK: domd: Add xt-optee-os with virtualization
This hack is adopted cherry-pick of
9e1b66e optee: use XT OP-TEE with virtualization enabled
from meta-xt-images.

Why it is here and why it is marked as hack?
Option 1.
Proper way to add xt-optee-os to this product is to add it to
domd-agl-cluster-drmo-platform. But we need to have this product
in 'fixed' state. Adding new commit to xt-images will result in
including the list of not needed commits.

Option 2.
To add bbappend to product's agl-cluster-demo.
But agl-cluster-demo was 'fast-and-dirty' solution that uses some
layers from image-weston. So we need to add new layer to agl-cluster
that will not conflict with existing in image-weston.
So this option 2 seems like complex hack.

Option 3.
Insert xt-optee-os as bbappend into image-weston.
It will not conflict with already present append in xt-images
for image-weston, because image-weston is not used in tu2019.
But it's layers are used in agl-cluster-demo.
And this is how we add xt-optee-os into tu2019 with minimal changes.

Sign-off of author of original commit is not included here
to avoid using of his name in this terrible solution.)

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>